requirements.txt
numpy==1.21.1
termcolor==1.1.0
rl-games==1.1.4
tensorboard==1.15.0
chumpy
vtk
numpy-stl
joblib
scikit-image
scikit-learn
scipy
matplotlib
torchgeometry
pillow
tqdm
lxml
opencv-python
wandb
ipdb
patchelf
termcolor
rl-games==1.1.4
pyyaml
mujoco-py<2.2,>=2.1
scikit-image
gym
git+https://github.com/ZhengyiLuo/smplx.git@master
human_body_prior
autograd
sklearn
chumpy